[INLONG-6207][Agent] Optimize the unit test TestSQLServerReader
